Split off `Cardano.DB.Sqlite.Migration.Old` module
