+++ lib/wallet/cardano-wallet.cabal
@@ -217,6 +217,7 @@ library
     Cardano.DB.Sqlite
     Cardano.DB.Sqlite.Delete
     Cardano.DB.Sqlite.ForeignKeys
+    Cardano.DB.Sqlite.Migration.Old
     Cardano.Pool.DB
     Cardano.Pool.DB.Log
     Cardano.Pool.DB.Model
