+++ lib/wallet/src/Cardano/Pool/DB/Sqlite.hs
@@ -31,18 +31,8 @@ module Cardano.Pool.DB.Sqlite
       -- * Helpers
     , handleConstraint
 
-      -- * Old-style Manual Migration
-    , ManualMigration (..)
-    , noManualMigration
-    , MigrationError (..)
-    , DBField (..)
-    , tableName
-    , fieldName
-    , fieldType
-
       -- * Logging
     , DBLog (..)
-    , foldMigrations
     ) where
 
 import Prelude
@@ -62,17 +52,17 @@ import Cardano.DB.Sqlite.ForeignKeys
     ( ForeignKeysSetting (..)
     , withForeignKeysDisabled
     )
-import Cardano.Wallet.DB.Migration
-    ( ErrWrongVersion (..)
-    )
-import Control.Lens
-    ( strict
-    , view
+import Cardano.DB.Sqlite.Migration.Old
+    ( DBField (..)
+    , ManualMigration (..)
+    , MatchMigrationError (..)
+    , MigrationError (..)
+    , fieldName
+    , tableName
     )
 import Control.Monad
     ( join
     , void
-    , when
     )
 import Control.Monad.IO.Unlift
     ( MonadUnliftIO (..)
@@ -106,42 +96,21 @@ import Data.Functor
     ( ($>)
     , (<&>)
     )
-import Data.List
-    ( isInfixOf
-    )
-import Data.Proxy
-    ( Proxy (..)
-    )
 import Data.Text
     ( Text
     )
 import Data.Text.Class
     ( ToText (..)
     )
-import Data.Text.Lazy.Builder
-    ( toLazyText
-    )
 import Data.Time.Clock
     ( NominalDiffTime
     )
-import Database.Persist.EntityDef
-    ( getEntityDBName
-    )
-import Database.Persist.Names
-    ( EntityNameDB (..)
-    , unFieldNameDB
-    )
 import Database.Persist.Sql
-    ( EntityField
-    , LogFunc
+    ( LogFunc
     , Migration
-    , PersistEntity (..)
     , PersistException
     , SqlPersistT
-    , SqlType (..)
     , close'
-    , fieldDB
-    , fieldSqlType
     , runMigrationUnsafeQuiet
     , runSqlConn
     )
@@ -154,8 +123,7 @@ import Database.Sqlite
     , SqliteException (SqliteException)
     )
 import Fmt
-    ( Buildable (..)
-    , fmt
+    ( fmt
     , ordinalF
     , (+|)
     , (+||)
@@ -175,9 +143,7 @@ import UnliftIO.Compat
     ( handleIf
     )
 import UnliftIO.Exception
-    ( Exception
-    , bracket
-    , bracket_
+    ( bracket
     , handleJust
     , tryJust
     )
@@ -187,7 +153,6 @@ import UnliftIO.MVar
     , withMVarMasked
     )
 
-import qualified Data.Aeson as Aeson
 import qualified Data.ByteString.Char8 as B8
 import qualified Data.Text as T
 import qualified Data.Text.Encoding as T
@@ -398,48 +363,8 @@ retryOnBusy tr timeout action =
 
 {-------------------------------------------------------------------------------
     Database migrations
-    old style and new style
+    old style
 -------------------------------------------------------------------------------}
-
--- | Error type for when migrations go wrong after opening a database.
-newtype MigrationError = MigrationError
-    {getMigrationErrorMessage :: Text}
-    deriving (Show, Eq, Generic, ToJSON)
-
-instance Exception MigrationError
-
-class Exception e => MatchMigrationError e where
-    -- | Exception predicate for migration errors.
-    matchMigrationError :: e -> Maybe MigrationError
-
-instance MatchMigrationError PersistException where
-    matchMigrationError e
-        | mark `isInfixOf` msg = Just $ MigrationError $ T.pack msg
-        | otherwise = Nothing
-      where
-        msg = show e
-        mark = "Database migration: manual intervention required."
-
-instance MatchMigrationError SqliteException where
-    matchMigrationError (SqliteException ErrorConstraint _ msg) =
-        Just $ MigrationError msg
-    matchMigrationError _ =
-        Nothing
-
-matchWrongVersionError :: ErrWrongVersion -> Maybe MigrationError
-matchWrongVersionError =
-    Just
-        . MigrationError
-        . view strict
-        . toLazyText
-        . build
-
--- | Encapsulates an old-style manual migration action
---   (or sequence of actions) to be
---   performed immediately after an SQL connection is initiated.
-newtype ManualMigration = ManualMigration
-    {executeManualMigration :: Sqlite.Connection -> IO ()}
-
 runAutoMigration
     :: Tracer IO DBLog
     -> Migration
@@ -479,65 +404,6 @@ runAllMigrations
 runAllMigrations tr fp old auto = runExceptT $ do
     ExceptT $ withDBHandle tr fp $ runManualOldMigrations tr old
     ExceptT $ withDBHandle tr fp $ runAutoMigration tr auto
-
-{-------------------------------------------------------------------------------
-    Database migration helpers
--------------------------------------------------------------------------------}
-
-noManualMigration :: ManualMigration
-noManualMigration = ManualMigration $ const $ pure ()
-
-foldMigrations :: [Sqlite.Connection -> IO ()] -> ManualMigration
-foldMigrations ms = ManualMigration $ \conn -> mapM_ ($ conn) ms
-
-data DBField where
-    DBField
-        :: forall record typ
-         . (PersistEntity record)
-        => EntityField record typ
-        -> DBField
-
-tableName :: DBField -> Text
-tableName (DBField (_ :: EntityField record typ)) =
-    unEntityNameDB $ getEntityDBName $ entityDef (Proxy @record)
-
-fieldName :: DBField -> Text
-fieldName (DBField field) =
-    unFieldNameDB $ fieldDB $ persistFieldDef field
-
-fieldType :: DBField -> Text
-fieldType (DBField field) =
-    showSqlType $ fieldSqlType $ persistFieldDef field
-
-showSqlType :: SqlType -> Text
-showSqlType = \case
-    SqlString -> "VARCHAR"
-    SqlInt32 -> "INTEGER"
-    SqlInt64 -> "INTEGER"
-    SqlReal -> "REAL"
-    SqlDay -> "DATE"
-    SqlTime -> "TIME"
-    SqlDayTime -> "TIMESTAMP"
-    SqlBlob -> "BLOB"
-    SqlBool -> "BOOLEAN"
-    SqlOther t -> t
-    SqlNumeric precision scale ->
-        T.concat
-            [ "NUMERIC("
-            , T.pack (show precision)
-            , ","
-            , T.pack (show scale)
-            , ")"
-            ]
-
-instance Show DBField where
-    show field = T.unpack (tableName field <> "." <> fieldName field)
-
-instance Eq DBField where
-    field0 == field1 = show field0 == show field1
-
-instance ToJSON DBField where
-    toJSON = Aeson.String . T.pack . show
 
 {-------------------------------------------------------------------------------
                                     Logging
